only gzip userdata in aws provider
